billogram_discount_code/app/requirements.txt
pandas
djangorestframework
pylint